Se refactorizan todos los id en Integer y se modela logica para BD en Company y Field.

+++ src/main/java/com/fulboapp/dao/IMatchDAO.java
@@ -38,7 +38,7 @@ public interface IMatchDAO {
    * @param matchID - Long del ID.
    * @return objeto MatchDTO o null si no hay coincidencias.
    */
-  MatchDTO findMatch(Long matchID);
+  MatchDTO findMatch(Integer matchID);
 
   /**
    * Busca los partidos guardados.
